Create code
In this project, you will prompt a user for a number of random numbers that they desire. (All numbers will be between 1 and 100.) You will then prompt the user for a 'block size'. For example, should the user respond with 1000 and 5, the program generates a file of 1000 random numbers (between 1 and 100) and will scan the file looking for which 5 consecutive numbers add to the greatest sum. The program will display the sum AND the starting position of the 5 numbers within the file.